pyproject.toml
[tool.poetry]
name = "jorevi"
version = "0.1.0"
description = "Version of remote job"
authors = ["xblzbjs <t710255295@gmail.com>"]
license = "BSD-3"

[tool.poetry.dependencies]
python = "^3.9"
python-slugify = "^6.1.1"
Pillow = "^9.1.0"
argon2-cffi = "^21.3.0"
redis = "^4.2.1"
celery = "^5.2.3"
flower = "^1.0.0"
phonenumbers = "^8.12.46"
feedparser = "^6.0.8"
django-environ = "^0.8.1"
django-model-utils = "^4.2.0"
django-import-export = "^2.8.0"
django-phonenumber-field = "^6.1.0"
sorl-thumbnail = "^12.8.0"
django-markdownx = "^4.0.0b1"
django-taggit = "^2.1.0"
django-filter = "^21.1"
django-mptt = "^0.13.4"
django-tinymce = "^3.4.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
drf-writable-nested = "^0.6.3"
django-grappelli = "^3.0.3"
django-filebrowser = "^4.0.1"
django-reversion = "^5.0.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
sentry-sdk = "^1.5.8"
djangorestframework-gis = "^0.18"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.6"
django-compressor = "^4.0"
django-allauth = "^0.50.0"
uvicorn = { extras = ["standard"], version = "^0.17.6" }
django-countries = "^7.3.2"
Django = "^4.0.5"

[tool.poetry.dev-dependencies]
Werkzeug = "2.0.3"
django-extensions = "^3.1.5"
ipdb = "^0.13.9"
psycopg2-binary = "^2.9.3"
watchgod = "^0.8.2"
mypy = "^0.942"
django-stubs = "^1.10.1"
pytest = "^7.1.1"
pytest-sugar = "^0.9.4"
djangorestframework-stubs = "^1.4.0"
mdgen = "^0.1.9"
Sphinx = "^4.5.0"
sphinx-autobuild = "^2021.3.14"
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"
coverage = "^6.3.2"
black = "^22.3.0"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"
pre-commit = "^2.17.0"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.2.4"
django-coverage-plugin = "^2.0.2"
pytest-django = "^4.5.2"
flake8-bugbear = "^22.3.23"
flake8-comprehensions = "^3.8.0"
flake8-simplify = "^0.19.2"
pytest-cov = "^3.0.0"
pep8-naming = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
